Fix user exiting bug

+++ websocketServerModule/src/sprout/clipcon/server/controller/Server.java
@@ -7,8 +7,10 @@ import java.util.Collections;
 import java.util.Date;
 import java.util.HashMap;
 import java.util.Iterator;
+import java.util.List;
 import java.util.Map;
 import java.util.Random;
+import java.util.Scanner;
 import java.util.Set;
 
 import sprout.clipcon.server.model.Group;
@@ -50,7 +52,7 @@ public class Server {
 	 */
 	public Group getGroupByPrimaryKey(String key) {
 
-		Set set = groups.keySet();
+		Set<String> set = groups.keySet();
 		Iterator<String> it = set.iterator();
 		while (it.hasNext()) {
 			System.out.print(it.next() + ", ");
@@ -96,7 +98,8 @@ public class Server {
 				for (int i = 0; i < tempFile.length; i++) {
 					if (tempFile[i].isFile()) {
 						tempFile[i].delete();
-					} else { // Recursive function
+					}
+					else { // Recursive function
 						deleteAllFilesInGroupDir(tempFile[i].getPath());
 					}
 					tempFile[i].delete();
@@ -108,7 +111,9 @@ public class Server {
 
 	public void removeGroup(Group group) {
 		Group removeGroup = groups.remove(group.getPrimaryKey());
-		deleteAllFilesInGroupDir(removeGroup.getPrimaryKey());
+		if (removeGroup != null) {
+			deleteAllFilesInGroupDir(removeGroup.getPrimaryKey());
+		}
 	}
 
 	/**
@@ -157,7 +162,8 @@ public class Server {
 				File file = fileList[i];
 				if (file.isFile()) {
 					System.out.println("File name = " + file.getPath());
-				} else if (file.isDirectory()) {
+				}
+				else if (file.isDirectory()) {
 					System.out.println("Dir name = " + file.getPath());
 					subDirList(file.getCanonicalPath().toString());
 				}
@@ -170,5 +176,48 @@ public class Server {
 		Date date = new Date();
 		SimpleDateFormat sdf = new SimpleDateFormat("yyyy-MM-dd, a hh:mm:ss");
 		System.out.println(sdf.format(date).toString());
+	}
+
+	class ServerStatusCheck extends Thread {
+
+		@Override
+		public void run() {
+			Scanner sc = new Scanner(System.in);
+			String input = null;
+			while (true) {
+				input = sc.nextLine();
+				switch (input) {
+				case "g":
+					System.out.println("┌──────────────SERVER STATUS");
+					System.out.println("├ Group Count: " + groups.size());
+					System.out.println("├ Group List...  ");
+					Set<String> set = groups.keySet();
+					Iterator<String> it = set.iterator();
+					while (it.hasNext()) {
+						System.out.println("│\t- " + it.next());
+					}
+					break;
+
+				default:
+					Group search = groups.get(input);
+					if (search == null) {
+						System.out.println("[SERVER] Group named \"" + input + "\" is not found.");
+					}
+					else {
+						List<String> list = search.getUserList();
+						System.out.println("┌──────────────GROUP INFO: " + input);
+						System.out.println("├ User Count: " + list.size());
+						System.out.println("├ User List...  ");
+						Iterator<String> listIt = list.iterator();
+						while (listIt.hasNext()) {
+							System.out.println("│\t- " + listIt.next());
+						}
+					}
+					break;
+				}
+				System.out.println("└───────────────────────────");
+				input = null;
+			}
+		}
 	}
 }
